// src/app/bridge/enums/account.enum.ts
export enum AccountIDs {
	GBM = 'gbm',
	POLYGON = 'plg',
	CONCORDIUM = 'ccd'
}

export enum AccountTitles {
	GBM = 'GBM',
	POLYGON = 'Polygon',
	CONCORDIUM = 'Concordium'
}

export enum AccountAliases {
	GBM = 'GBM',
	POLYGON = 'WGBM',
	CONCORDIUM = 'WGBM'
}

export enum AccountIcons {
	GBM = 'gbmCoin',
	POLYGON = 'matic',
	CONCORDIUM = 'eth'
}